almost finished

+++ DBHandler.cs
@@ -219,22 +219,28 @@ namespace myKEP
             } else
             {
                 NewUserEntry(user, kepDB);
-                MessageBox.Show("Ο χρήστης με αριθμό Ταυτότητας" + user.AT + "καταχωρίστηκε");
+                MessageBox.Show("Ο χρήστης με αριθμό Ταυτότητας" + user.AT + " καταχωρίστηκε");
             }
         }
 
         public void DeleteUser(string AT)
         {
-            SqlConnection kepDB = new SqlConnection(ConnString);
-            kepDB.Open();
-            string qDelUser = "DELETE FROM Users WHERE AT=@Parameter";
-            SqlCommand command = new SqlCommand(qDelUser, kepDB);
-            command.Parameters.AddWithValue("@Parameter", AT);
-            command.ExecuteNonQuery();
-            string qDelReqs = "DELETE FROM Requests WHERE AT=@Parameter";
-            command.CommandText = qDelReqs;
-            command.ExecuteNonQuery();
-            MessageBox.Show("Ο χρήστης με Αριθμό Ταυτότητας " + AT + " έχει διαγραφεί", "Message");
+            if (UserExists(AT))
+            {
+                SqlConnection kepDB = new SqlConnection(ConnString);
+                kepDB.Open();
+                string qDelUser = "DELETE FROM Users WHERE AT=@Parameter";
+                SqlCommand command = new SqlCommand(qDelUser, kepDB);
+                command.Parameters.AddWithValue("@Parameter", AT);
+                command.ExecuteNonQuery();
+                string qDelReqs = "DELETE FROM Requests WHERE AT=@Parameter";
+                command.CommandText = qDelReqs;
+                command.ExecuteNonQuery();
+                MessageBox.Show("Ο χρήστης με Αριθμό Ταυτότητας " + AT + " έχει διαγραφεί", "Message");
+            } else
+            {
+                MessageBox.Show("Δεν υπάρχει χρήστης με αυτό τον κωδικό", "Message");
+            }
         }
 
         public void DeleteRequest(string reqID)
@@ -244,23 +250,24 @@ namespace myKEP
             string qGetReqAT = "SELECT AT FROM Requests WHERE reqID = @Parameter1"; //Get the AT of the request's User, to reduce their reqNo
             SqlCommand command = new SqlCommand(qGetReqAT, kepDB);
             command.Parameters.AddWithValue("@Parameter1", reqID);
-            SqlDataReader reader = command.ExecuteReader();
+            SqlDataReader reader1 = command.ExecuteReader();
 
-            if (reader.HasRows)
+            if (reader1.HasRows)
             {
-                if (reader.Read())
+                if (reader1.Read())
                 {
-                    string AT = reader.GetString(0);
+                    string AT = reader1.GetString(0);
 
                     string qGetReqNo = "SELECT reqNo FROM Users WHERE AT = @Parameter2";    //Get reqNo and reduce it
                     command.CommandText = qGetReqNo;
                     command.Parameters.AddWithValue("@Parameter2", AT);
-                    reader = command.ExecuteReader();
-                    if (reader.Read())
+                    reader1.Close();
+                    SqlDataReader reader2 = command.ExecuteReader();
+                    if (reader2.Read())
                     {
-                        int reqNo = reader.GetInt32(0);
+                        int reqNo = reader2.GetInt32(0);
                         reqNo += -1;
-                        reader.Close();
+                        reader2.Close();
 
                         string qUpdateReqNo = "UPDATE Users SET reqNo = @Par1 WHERE AT = @Par2";    //Update reqNo in User
                         command.CommandText = qUpdateReqNo;
@@ -272,7 +279,7 @@ namespace myKEP
                         command.CommandText = qDelReq;
                         command.ExecuteNonQuery();
 
-                        MessageBox.Show("Η αίτηση με κωδικό " + reqID + "έχει διαγραφεί", "Message");
+                        MessageBox.Show("Η αίτηση με κωδικό " + reqID + " έχει διαγραφεί", "Message");
                     }
 
                 }
